requirements.txt
cycler==0.10.0
Django==3.1.7
django-crispy-forms==1.11.2
django-filter==2.4.0
MarkupSafe==1.1.1
psycopg2==2.8.6
python-dateutil==2.8.1
python-decouple==3.4
regex==2021.4.4
requests==2.24.0
urllib3==1.25.10
virtualenv==20.4.4
virtualenv-clone==0.5.4
